fix: split variable class
