+++ octopus/dsl/__init__.py
@@ -2,6 +2,6 @@
 DSL module for test configuration and parsing
 """
 
-from . import dsl_config, dsl_validator, parser
+from . import dsl_config, dsl_validator, parser, variable
 
-__all__ = ["dsl_config", "parser", "dsl_validator"]
+__all__ = ["dsl_config", "parser", "dsl_validator", "variable"]
